requirements.txt
colorama==0.2.5
configobj==4.7.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
duplicity==0.6.23
futures==3.0.5
html5lib==0.999
httplib2==0.8
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lxml==3.3.3
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
PAM==0.4.2
paramiko==1.10.1
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyOpenSSL==0.13
pyserial==2.6
pysmbc==1.0.14.1
python-swiftclient==3.0.0
pyxdg==0.25
reportlab==3.0
requests==2.2.1
SecretStorage==2.0.0
sessioninstaller==0.0.0
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
system-service==0.1.6
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3==1.7.1
wadllib==1.3.2
wheel==0.24.0
xdiagnose===3.6.3build2
zope.interface==4.0.5
